requirements.txt
numpy>=1.24,<3
pandas>=2.0
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
rasterio>=1.3
Pillow>=10.0
PyYAML>=6.0
requests>=2.31
earthengine-api>=1.6
